% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpDecomposition.R
\name{cpDecomposition}
\alias{cpDecomposition}
\title{Canonical Polyadic Decomposition}
\usage{
cpDecomposition(tnsr, num_components = NULL, max_iter = 25, tol = 1e-05)
}
\arguments{
\item{tnsr}{Tensor with K modes}

\item{num_components}{the number of rank-1 K-Tensors to use in approximation}

\item{max_iter}{maximum number of iterations if error stays above tol}

\item{tol}{relative Frobenius norm error tolerance}
}
\value{
a list containing decomposition results
}
\description{
CP decomposition of a tensor using Alternating Least Squares
}
