% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dRegulationCpp}
\alias{dRegulationCpp}
\title{Differential regulation analysis with C++ acceleration}
\usage{
dRegulationCpp(manifoldOutputSEXP, geneNames, gKO)
}
\arguments{
\item{geneNames}{Vector of gene names}

\item{gKO}{Name of knocked out gene}

\item{manifoldOutput}{Matrix with WT and KO gene embeddings (2*nGenes x d)}
}
\value{
DataFrame with differential regulation statistics
}
\description{
Differential regulation analysis with C++ acceleration
}
