% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knockoutGeneCpp}
\alias{knockoutGeneCpp}
\title{Perform virtual knockout on network}
\usage{
knockoutGeneCpp(networkSEXP, geneIdx)
}
\arguments{
\item{geneIdx}{Index of gene to knockout (0-based)}

\item{network}{Gene regulatory network matrix}
}
\value{
Network with gene knocked out
}
\description{
Perform virtual knockout on network
}
