% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{makeNetworksCpp}
\alias{makeNetworksCpp}
\title{Build multiple gene regulatory networks with C++ acceleration}
\usage{
makeNetworksCpp(
  countMatrixSEXP,
  nNet = 10L,
  nCells = 500L,
  nComp = 3L,
  q = 0.9,
  scaleScores = TRUE,
  symmetric = FALSE,
  nThreads = 0L
)
}
\arguments{
\item{nNet}{Number of networks to generate}

\item{nCells}{Number of cells to subsample for each network}

\item{nComp}{Number of principal components}

\item{q}{Quantile threshold for edge filtering}

\item{scaleScores}{Whether to scale network weights to [-1, 1]}

\item{symmetric}{Whether to make networks symmetric}

\item{nThreads}{Number of threads (0 = auto)}

\item{countMatrix}{Gene expression matrix (genes x cells)}
}
\value{
List of networks
}
\description{
Build multiple gene regulatory networks with C++ acceleration
}
