% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifoldAlignment.R
\name{manifoldAlignment}
\alias{manifoldAlignment}
\title{Performs non-linear manifold alignment of two gene regulatory networks}
\usage{
manifoldAlignment(X, Y, d = 30, nCores = parallel::detectCores())
}
\arguments{
\item{X}{A gene regulatory network}

\item{Y}{A gene regulatory network}

\item{d}{The dimension of the low-dimensional feature space}

\item{nCores}{An integer value. Defines the number of cores to be used (currently not used, kept for compatibility)}
}
\value{
A low-dimensional projection for the two gene regulatory networks used as input
}
\description{
Build comparable low-dimensional features for two weight-averaged denoised
single-cell gene regulatory networks using non-linear network embedding
}
