% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sparseToDenseTranspose}
\alias{sparseToDenseTranspose}
\title{Convert sparse matrix to dense and transpose}
\usage{
sparseToDenseTranspose(XSEXP)
}
\arguments{
\item{X}{Sparse matrix}
}
\value{
Dense transposed matrix
}
\description{
Convert sparse matrix to dense and transpose
}
