% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strictDirection.R
\name{strictDirection}
\alias{strictDirection}
\title{Apply directional penalty to network}
\usage{
strictDirection(X, lambda = 1)
}
\arguments{
\item{X}{Network matrix}

\item{lambda}{Penalty parameter between 0 and 1}
}
\value{
Modified network matrix
}
\description{
Applies lambda penalty to enforce edge directionality
}
