% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{strictDirectionCpp}
\alias{strictDirectionCpp}
\title{Enhance network directionality with C++ acceleration}
\usage{
strictDirectionCpp(XSEXP, lambda = 1)
}
\arguments{
\item{lambda}{Directionality parameter (0 to 1)}

\item{X}{Network adjacency matrix}
}
\value{
Enhanced network matrix
}
\description{
Enhance network directionality with C++ acceleration
}
