% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorDecomposition.R
\name{tensorDecomposition}
\alias{tensorDecomposition}
\title{Performs CANDECOMP/PARAFAC (CP) Tensor Decomposition}
\usage{
tensorDecomposition(
  xList,
  yList = NULL,
  nDecimal = 1,
  K = 5,
  maxError = 1e-05,
  maxIter = 1000
)
}
\arguments{
\item{xList}{A list of gene regulatory networks}

\item{yList}{Optional. A list of gene regulatory networks}

\item{nDecimal}{An integer value indicating the number of decimal places to be used}

\item{K}{The number of rank-one tensors used to approximate the data}

\item{maxError}{A decimal value between 0 and 1. Defines the relative Frobenius norm error tolerance}

\item{maxIter}{An integer value. Defines the maximum number of iterations}
}
\value{
A list of weight-averaged denoised gene regulatory networks
}
\description{
Generate weight-averaged denoised gene regulatory networks using CANDECOMP/PARAFAC (CP) Tensor Decomposition
}
