% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tensorDecompositionCpp}
\alias{tensorDecompositionCpp}
\title{CP Tensor Decomposition with C++ acceleration}
\usage{
tensorDecompositionCpp(
  networkList,
  K = 3L,
  maxIter = 1000L,
  maxError = 1e-05,
  nDecimal = 3L
)
}
\arguments{
\item{networkList}{List of network matrices}

\item{K}{Rank of decomposition}

\item{maxIter}{Maximum iterations}

\item{maxError}{Convergence threshold}

\item{nDecimal}{Decimal places for output}
}
\value{
List with reconstructed network and factors
}
\description{
CP Tensor Decomposition with C++ acceleration
}
