% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRegulation.R
\name{dRegulation}
\alias{dRegulation}
\title{Differential regulation analysis}
\usage{
dRegulation(manifoldOutput, gKO)
}
\arguments{
\item{manifoldOutput}{Manifold alignment output}

\item{gKO}{Knocked out gene name}
}
\value{
Data frame with differential regulation results
}
\description{
Computes differential regulation between aligned manifolds
}
