% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scQC.R
\name{scQC}
\alias{scQC}
\title{Single-cell quality control}
\usage{
scQC(X, mtThreshold = 0.1, minLSize = 1000)
}
\arguments{
\item{X}{A count matrix or Seurat object}

\item{mtThreshold}{Maximum mitochondrial read ratio threshold}

\item{minLSize}{Minimum library size threshold}
}
\value{
Filtered count matrix
}
\description{
Performs quality control filtering on single-cell RNA-seq data
}
