% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTenifoldKNK.R
\name{scTenifoldKNK}
\alias{scTenifoldKNK}
\title{scTenifoldKNK}
\usage{
scTenifoldKNK(countMatrix, gKO = NULL)
}
\arguments{
\item{countMatrix}{countMatrix}

\item{gKO}{gKO}
}
\description{
Predict gene perturbations
}
\author{
Daniel Osorio <dcosorioh@tamu.edu>
}
