% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTenifoldKNK.R
\name{scTenifoldKNK}
\alias{scTenifoldKNK}
\title{scTenifoldKNK}
\usage{
scTenifoldKNK(countMatrix, gKO = NULL, qc_mtThreshold = 0.1,
  qc_minLSize = 1000, nc_nNet = 10, nc_nCells = 500, nc_nComp = 3,
  nc_scaleScores = TRUE, nc_symmetric = FALSE, nc_q = 0.9,
  td_K = 3, td_maxIter = 1000, td_maxError = 1e-05, ma_nDim = 2)
}
\arguments{
\item{countMatrix}{countMatrix}

\item{gKO}{gKO}

\item{qc_mtThreshold}{A decimal value between 0 and 1. Defines the maximum ratio of mitochondrial reads (mithocondrial reads / library size) present in a cell to be included in the analysis. It's computed using the symbol genes starting with 'MT-' non-case sensitive.}

\item{qc_minLSize}{An integer value. Defines the minimum library size required for a cell to be included in the analysis.}

\item{nc_nNet}{An integer value. The number of networks based on principal components regression to generate.}

\item{nc_nCells}{An integer value. The number of cells to subsample each time to generate a network.}

\item{nc_nComp}{An integer value. The number of principal components in PCA to generate the networks. Should be greater than 2 and lower than the total number of genes.}

\item{nc_scaleScores}{A boolean value (TRUE/FALSE), if TRUE, the weights will be normalized such that the maximum absolute value is 1.}

\item{nc_symmetric}{A boolean value (TRUE/FALSE), if TRUE, the weights matrix returned will be symmetric.}

\item{nc_q}{A decimal value between 0 and 1. Defines the cut-off threshold of top q\% relationships to be returned.}

\item{td_K}{An integer value. Defines the number of rank-one tensors used to approximate the data using CANDECOMP/PARAFAC (CP) Tensor Decomposition.}

\item{td_maxIter}{An integer value. Defines the maximum number of iterations if error stay above \code{td_maxError}.}

\item{td_maxError}{A decimal value between 0 and 1. Defines the relative Frobenius norm error tolerance.}

\item{ma_nDim}{An integer value. Defines the number of dimensions of the low-dimensional feature space to be returned from the non-linear manifold alignment.}
}
\description{
Predict gene perturbations
}
\author{
Daniel Osorio <dcosorioh@tamu.edu>
}
